Updated code to handle realisation numbers greater than 1 and remove unused...

Closes #320 and #321

See merge request APS/GUI!366

+++ aps/algorithms/truncation_rules/Trunc2D_Base_xml.py
@@ -1586,7 +1586,7 @@ Background facies:
                 facies_code = self._faciesCode[fIndx]
                 fIndx_vector = np.ones(len(alpha_coord_vectors[:, self._alphaIndxList[0]]), dtype=int) * fIndx
                 faciesCode_vector = np.ones(len(alpha_coord_vectors[:, self._alphaIndxList[0]]),
-                                            dtype=int) * facies_code
+                                            dtype=np.uint8) * facies_code
                 return faciesCode_vector, fIndx_vector
 
         x_coordinates = alpha_coord_vectors[:, self._alphaIndxList[0]]
@@ -1595,7 +1595,7 @@ Background facies:
         # Input is facies polygons for truncation rules and coordinates in alpha space for
         # for a set of points saved in vectors of length equal to the set of points.
         # Check in which polygon the points are located and thereby the facies
-        faciesCode_vector = np.ones(len(x_coordinates), dtype=int) * (-1)
+        faciesCode_vector = np.ones(len(x_coordinates), dtype=np.uint8) * (-1)
         for poly_number in range(self.num_polygons):
             polygon = self._faciesPolygons[poly_number]
             selected = (polygon_number_all_vector == -1)
